requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
executing==2.2.0
frozenlist==1.6.0
icecream==2.1.4
idna==3.10
magic-filter==1.0.12
multidict==6.4.4
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.20.0
